+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Added
 - Added support node.js v18.20.0
+### Fixed
+- Bug with line breaks for the schema content
+
 
 ## [0.28.0] - 2024-10-25
 ### Added